requirements.txt
selenium
webdriver-manager
beautifulsoup4
pandas
snowflake-connector-python
python-dotenv